Cast away integer comparison

+++ Geo/GEdge.cpp
@@ -752,7 +752,7 @@ bool GEdge::reorder(const int elementType, const std::vector<int> &ordering)
 
   for(std::vector<int>::const_iterator it = ordering.begin();
       it != ordering.end(); ++it) {
-    if(*it < 0 || *it >= lines.size()) return false;
+    if(*it < 0 || *it >= static_cast<int>(lines.size())) return false;
   }
 
   std::vector<MLine *> newLinesOrder(lines.size());
